browser_research/requirements.txt
# Core dependencies
asyncio>=3.4.3
pyyaml>=6.0
aiohttp>=3.8.3
beautifulsoup4>=4.11.1

# Browser automation
playwright>=1.30.0

# Web interface
fastapi>=0.92.0
uvicorn>=0.20.0
jinja2>=3.1.2
pydantic>=1.10.5

# Additional utilities
python-multipart>=0.0.5
websockets>=10.4
python-dateutil>=2.8.2
markdown>=3.4.1
